Move fingerprint creation into fingerprint module and add brand
